Move form configs out of preprocessor
