+++ common/src/python/configs/BUILD
@@ -1,0 +1,1 @@
+python_sources()
